+++ proto/client-api/src/lib.rs
@@ -186,23 +186,38 @@ pub enum Command {
     ResetMe,
 }
 
-/// Reason of disconnecting client from the server.
+/// Reason of disconnecting Web Client from Media Server.
 #[derive(Debug, Deserialize, Serialize)]
-pub enum RpcConnectionCloseReason {
-    /// Client session was finished on the server side.
+pub enum CloseReason {
+    /// Client session was finished on a server side.
     Finished,
 
-    /// Old connection was closed due to client reconnection.
-    NewConnection,
+    /// Old connection was closed due to a client reconnection.
+    Reconnected,
+
+    /// Connection has been inactive for a while and thus considered idle
+    /// by a server.
+    Idle,
+
+    /// Establishing of connection with a server was rejected on server side.
+    ///
+    /// Most likely because of incorrect Member credentials.
+    Rejected,
+
+    /// Server internal error has occurred while connecting.
+    ///
+    /// This close reason is similar to 500 HTTP status code.
+    InternalError,
 }
 
-/// Description which will be sent in [Close] WebSocket frame.
+/// Description which is sent in [Close] WebSocket frame from Media Server
+/// to Web Client.
 ///
 /// [Close]: https://tools.ietf.org/html/rfc6455#section-5.5.1
 #[derive(Constructor, Debug, Deserialize, Serialize)]
 pub struct CloseDescription {
-    /// Reason of why connection was closed.
-    pub reason: RpcConnectionCloseReason,
+    /// Reason of why WebSocket connection has been closed.
+    pub reason: CloseReason,
 }
 
 /// WebSocket message from Medea to Jason.
@@ -528,8 +543,8 @@ mod test {
                 \"command\":\"MakeSdpOffer\",\
                 \"data\":{\
                     \"peer_id\":77,\
-	                \"sdp_offer\":\"offer\",\
-	                \"mids\":{\"0\":\"1\"}\
+                    \"sdp_offer\":\"offer\",\
+                    \"mids\":{\"0\":\"1\"}\
                 }\
             }";