+++ src/api/client/rpc_connection.rs
@@ -46,9 +46,15 @@ pub struct EventMessage(Event);
 ///
 /// [`Member`]: crate::signalling::elements::member::Member
 pub trait RpcConnection: fmt::Debug + Send {
-    /// Closes [`RpcConnection`].
+    /// Closes [`RpcConnection`] and sends [`CloseDescription`] to the client
+    /// (in WebSocket implementation description will be sent in a [Close]
+    /// frame).
+    ///
     /// No [`RpcConnectionClosed`] signals should be emitted.
-    /// Always returns success.
+    ///
+    /// Always succeeds.
+    ///
+    /// [Close]: https://tools.ietf.org/html/rfc6455#section-5.5.1
     fn close(
         &mut self,
         close_description: CloseDescription,